Use MUnit in ByteBufferInputStreamSuite
